Renamed tex folder and plots, added bin folder finished P7
The makefile now moves the .o files into the bin folder. Data will now be written to the data folder. This includes plots and text files. The LaTeX file now gets the plots from the data folder. I also edited and renamed the python script to plot data. The solution to problem 7 is mirrored on the y-axis. Must have some sign error somewhere