Changed algorithm and n = 10^6 is done
Made a file runtimes.txt where we can fill in the runtimes for n = 10 up to n = 10^6 for task 10. Plotted again for all these n, and zoomed in and saved for comparison. Made some new additions to the todo list on main.cpp. Changed the find_v function to take in vectors a, b and c, as problem 6 asks for this. Moved the original algorithm to find_v_general in case this was correct for the general algorithm